+++ crates/server/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod api;
+pub mod auth;
 pub mod components;
 pub mod configuration;
 pub mod domain;
